+++ CHANGELOG.md
@@ -204,3 +204,10 @@
 * EPMRPP-31447 - WS: "ActionType"=analyze_item for action that was done by user
 * Fix issue with incorrect importing of some junit files
 * EPMRPP-31904 - WS: Jobs do not work properly
+
+## 4.x.x
+##### Released: XXX XX, 2018
+
+### New Features
+
+* EPMRPP-31780 - Set the name of analyzer which made the changes